Retrieve the status of all Taxa according to a Status Type

Retrieve the status of all Taxa according to a Status Type

rt_interactions_find_type(interaction_id = NULL, page = 1,
  size = 2000)

Arguments

interaction_id
character(1)
The identifier of status (see rt_interactions_type() for the list of status types)

page
integer(1); default = 1
The page number returned

size
integer(1); default = 2000
The number of lines returned per page (maximum = 2000)

See also

Examples

rt_interactions_find_type(interaction_id = "GALLES")

#> New names:
#> * `_links.self.href` -> ._links.self.href

#> # A tibble: 554 x 27
#>    taxon.id taxon.scientifi… taxon.fullNameH… taxon.reference… taxon.parentId
#>       <int> <chr>            <chr>                       <int> <lgl>         
#>  1   221353 Acalitus brevit… <i>Acalitus bre…           221353 NA            
#>  2   221354 Acalitus calyco… <i>Acalitus cal…           221354 NA            
#>  3   848629 Acalitus rudis … <i>Acalitus rud…           221355 NA            
#>  4   221359 Acalitus stenas… <i>Acalitus ste…           221359 NA            
#>  5   848633 Aceria acerisca… <i>Aceria aceri…           848633 NA            
#>  6   848635 Aceria acerismo… <i>Aceria aceri…           848635 NA            
#>  7   221251 Aceria achilleae <i>Aceria achil…           221251 NA            
#>  8   848637 Aceria alpestris <i>Aceria alpes…           848637 NA            
#>  9   221254 Aceria anceps    <i>Aceria ancep…           221254 NA            
#> 10   848639 Aceria aroniae   <i>Aceria aroni…           848639 NA            
#> # … with 544 more rows, and 22 more variables: taxon.referenceNameHtml <chr>,
#> #   relationId <chr>, relationName <chr>, specificity <chr>, inherited <lgl>,
#> #   target.id <int>, target.scientificName <chr>, target.fullNameHtml <chr>,
#> #   target.referenceId <int>, target.parentId <lgl>,
#> #   target.referenceNameHtml <chr>, source.id <int>,
#> #   source.shortCitation <lgl>, source.year <lgl>, source.fullCitation <chr>,
#> #   source.url <chr>, source.doi <lgl>, source.doiUri <lgl>,
#> #   source.zooBankId <lgl>, source.zooBankUri <lgl>, source.abstract <lgl>,
#> #   ._links.self.href <chr>