Search Interactions based on a list of criteria

Search Interactions based on a list of criteria

rt_interactions_search(
  id = NULL,
  interaction_id = NULL,
  opgroup_id = NULL,
  page = 1,
  size = 2000
)

Arguments

id

integer(1)
an integer giving the id of the taxon in TaxRef (also called cdNom)

interaction_id

character(1), default = NULL
A character identifying the type of interactions

opgroup_id

numeric(1), default = NULL
the id of an operational group referenced in TaxRef

page

integer(1); default = 1
The page number returned

size

integer(1); default = 5000
The number of lines returned per page (maximum = 5000)

Examples

# Interactors with 'Ophrys apifera' (id = 110335)
rt_interactions_search(id = 110335)
#> New names:
#> • `_links.taxon.href` -> `._links.taxon.href`
#> • `_links.self.href` -> `._links.self.href`
#> • `_links.target.href` -> `._links.target.href`
#> • `_links.specificity.href` -> `._links.specificity.href`
#> # A tibble: 1 × 34
#>   taxonSex taxonStage taxon.id taxon.scientificName taxon.fullNameHtml          
#>   <lgl>    <lgl>         <int> <chr>                <chr>                       
#> 1 NA       NA           241554 Exosoma lusitanicum  <i>Exosoma lusitanicum</i> …
#> # … with 29 more variables: taxon.referenceId <int>, taxon.parentId <int>,
#> #   taxon.referenceNameHtml <chr>, ._links.taxon.href <chr>, id <int>,
#> #   relationId <chr>, relationName <chr>, specificity <chr>, inherited <lgl>,
#> #   targetStage <lgl>, target.id <int>, target.scientificName <chr>,
#> #   target.fullNameHtml <chr>, target.referenceId <int>, target.parentId <int>,
#> #   target.referenceNameHtml <chr>, source.id <int>,
#> #   source.shortCitation <lgl>, source.year <lgl>, source.fullCitation <chr>, …