Retrieve the status of all Taxa according to a Status Type

Retrieve the status of all Taxa according to a Status Type

rt_status_find_type(status_id = NULL, page = 1, size = 2000)

Arguments

status_id

character(1)
The identifier of status (see rt_status_type() for the list of status types)

page

integer(1); default = 1
The page number returned

size

integer(1); default = 2000
The number of lines returned per page (maximum = 2000)

See also

Examples

# Find all taxa concerned by Habitat Directive referenced in TaxRef
rt_status_find_type("DH")
#> # A tibble: 835 × 22
#>        id scientificName      fullNameHtml referenceId parentId referenceNameHt…
#>     <int> <chr>               <chr>              <int>    <int> <chr>           
#>  1 717038 Abies nebrodensis   <i>Abies ne…      717038   188665 <i>Abies nebrod…
#>  2 717038 Abies nebrodensis   <i>Abies ne…      717038   188665 <i>Abies nebrod…
#>  3 347916 Acipenser oxyrinch… <i>Acipense…      347916   188758 <i>Acipenser ox…
#>  4  66775 Acipenser sturio    <i>Acipense…       66775   188758 <i>Acipenser st…
#>  5  66775 Acipenser sturio    <i>Acipense…       66775   188758 <i>Acipenser st…
#>  6 610608 Acis nicaeensis     <i>Acis nic…      610608   606844 <i>Acis nicaeen…
#>  7 610608 Acis nicaeensis     <i>Acis nic…      610608   606844 <i>Acis nicaeen…
#>  8 130784 Aconitum napellus … <i>Aconitum…      130784    80037 <i>Aconitum nap…
#>  9 130784 Aconitum napellus … <i>Aconitum…      130784    80037 <i>Aconitum nap…
#> 10  81197 Aldrovanda vesicul… <i>Aldrovan…       81197   188945 <i>Aldrovanda v…
#> # … with 825 more rows, and 16 more variables: href <chr>,
#> #   statusTypeName <chr>, statusTypeGroup <chr>, statusCode <chr>,
#> #   statusName <chr>, statusRemarks <chr>, locationId <chr>,
#> #   locationName <chr>, locationAdminLevel <chr>, sourceId <int>, source <chr>,
#> #   requestedLocationId <lgl>, listedName <lgl>, statusType.href <chr>,
#> #   location.href <chr>, bibliographicResource.href <chr>