Search for a taxon information

Search for a taxon information

rt_taxa_search(
  id = NULL,
  sciname = NULL,
  fr_name = NULL,
  en_name = NULL,
  rank = NULL,
  territories = NULL,
  domain = NULL,
  habitats = NULL,
  vernacular = NULL,
  version = NULL,
  page = 1,
  size = 5000
)

Arguments

id

NULL or integer(1+) [default = NULL]
One or more integer giving the id(s) of taxa in TAXREF (also called cdNom)

sciname

NULL or character(1+) [default = NULL]
One or more scientific name of searched taxa

fr_name

NULL or character(1+) [default = NULL]
One or more common name(s) in French of searched taxa

en_name

NULL or character(1+) [default = NULL]
One or more common name(s) in English of searched taxa

rank

NULL or character(1+) [default = NULL]
One of more rank of the searched taxa (see rt_taxo_ranks() for the list of all ranks available):

Dumm= Domain

KD= Kingdom

PH= Phylum

CL= Class

OR= Order

FM= Family

SBFM= Subfamily

TR= Tribe

GN= Genus

AGES= Aggregate

ES= Species

SSES= Subspecies

NAT= Natio

VAR= Variety

SVAR= Sub-variety

FO= Form

SSFO= Sub-form

RACE= Race

CAR= Cultivar

AB= Abberatio

territories

NULL or character(1+) [default = NULL]
One or more territories where species is present:

"fr" = Mainland France

"gf" = French Guiana

"gua" = Guadeloupe

"mar" = Martinique

"sm" = Saint Martin

"sb" = Saint Barthélemy

"spm" = Saint Pierre and Miquelon

"epa" = Scattered Islands in the Indian Ocean

"may" = Mayotte

"reu" = Réunion

"sa" = French Southern and Antarctic Lands

"ta" = Adélie Land

"nc" = New Caledonia

"wf" = Wallis and Futuna

"pf" = French Polynesia

"cli" = Clipperton Island

domain

NULL or character(1) [default = NULL]
The domain where the species is found:

"marin" = Marine

"continental" = Continental

habitats

NULL or integer(1+) [default = NULL]
One or more habitats where the species is found (see rt_habitats() for a list of all habitats):

1 = Marine

2 = Freshwater

3 = Terrestrial

4 = Marine and freshwater

5 = Marine and Terrestrial

6 = Brackish water

7 = Continental (terrestrial and/or freshwater)

8 = Continental (terrestrial and freshwater)

vernacular

NULL or character(1+) [default = NULL]
Name of a vernacular group (see the list using rt_vernacular_groups())

version

NULL or character(1) [default = NULL]
A string indicating which version of TAXREF should be queried. Should be of the form "2.0" up to "15.0". You can check the available TAXREF versions with rt_taxref_versions(). If NULL uses the latest version of TAXREF available.

page

integer(1) [default = 1]
The page number returned

size

integer(1) [default = 5000]
The number of lines returned per page (maximum = 5000)

See also

Examples

if (FALSE) rt_taxa_search(sciname = "Bradypus") # \dontrun{}