Retrieve the status of all Taxa according to a Status Type

Retrieve the status of all Taxa according to a Status Type

rt_interactions_find_type(interaction_id = NULL, page = 1, size = 2000)

Arguments

interaction_id

character(1)
The identifier of status (see rt_interactions_type() for the list of status types)

page

integer(1); default = 1
The page number returned

size

integer(1); default = 2000
The number of lines returned per page (maximum = 2000)

See also

Examples

rt_interactions_find_type(interaction_id = "GALLES")
#> New names:
#> • `_links.taxon.href` -> `._links.taxon.href`
#> • `_links.taxonStage.href` -> `._links.taxonStage.href`
#> • `_links.self.href` -> `._links.self.href`
#> • `_links.target.href` -> `._links.target.href`
#> • `_links.specificity.href` -> `._links.specificity.href`
#> • `_links.targetStage.href` -> `._links.targetStage.href`
#> # A tibble: 2,000 × 36
#>    taxonSex taxonStage taxon.id taxon.scientificName    taxon.fullNameHtml      
#>    <lgl>    <chr>         <int> <chr>                   <chr>                   
#>  1 NA       NA           221353 Acalitus brevitarsus    <i>Acalitus brevitarsus…
#>  2 NA       NA           221353 Acalitus brevitarsus    <i>Acalitus brevitarsus…
#>  3 NA       NA           221353 Acalitus brevitarsus    <i>Acalitus brevitarsus…
#>  4 NA       NA           221353 Acalitus brevitarsus    <i>Acalitus brevitarsus…
#>  5 NA       NA           221353 Acalitus brevitarsus    <i>Acalitus brevitarsus…
#>  6 NA       NA           221354 Acalitus calycophthirus <i>Acalitus calycophthi…
#>  7 NA       NA           638051 Acalitus capparidis     <i>Acalitus capparidis<…
#>  8 NA       NA           720162 Acalitus granulatus     <i>Acalitus granulatus<…
#>  9 NA       NA           221358 Acalitus rudis          <i>Acalitus rudis</i> (…
#> 10 NA       NA           221358 Acalitus rudis          <i>Acalitus rudis</i> (…
#> # … with 1,990 more rows, and 31 more variables: taxon.referenceId <int>,
#> #   taxon.parentId <int>, taxon.referenceNameHtml <chr>,
#> #   ._links.taxon.href <chr>, ._links.taxonStage.href <chr>, id <int>,
#> #   relationId <chr>, relationName <chr>, specificity <chr>, inherited <lgl>,
#> #   targetStage <chr>, target.id <int>, target.scientificName <chr>,
#> #   target.fullNameHtml <chr>, target.referenceId <int>, target.parentId <int>,
#> #   target.referenceNameHtml <chr>, source.id <int>, …